--- src/app/(public)/profile/[id]/page.tsx
@@ -1,10 +1,0 @@
-import React from "react";
-import ProfileForm from "./profileform";
-
-const page = async () => {
-  return (
-    <ProfileForm/>
-  );
-};
-
-export default page;
